requirements.txt
psycopg[binary]==3.3.3
python-dotenv